+++ src/main/java/se/mow_e/controllers/AuthController.java
@@ -3,17 +3,13 @@ package se.mow_e.controllers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import se.mow_e.models.LoginRequest;
-import se.mow_e.services.JwtService;
+import se.mow_e.services.AuthService;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -23,30 +19,21 @@ import java.util.Map;
 public class AuthController {
 
     @Autowired
-    private AuthenticationManager authenticationManager;
-
-    @Autowired
-    private JwtService jwtService;
+    private AuthService authService;
 
     @PostMapping("/login")
     public ResponseEntity<?> authenticateUser(@RequestBody LoginRequest loginRequest) {
         Map<String, Object> response = new HashMap<>();
 
         try {
-            // Authenticate user
-            Authentication authentication = authenticationManager.authenticate(
-                    new UsernamePasswordAuthenticationToken(loginRequest.getUsername(), loginRequest.getPassword())
-            );
-
-            // Generate JWT token
-            UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-            String jwtToken = jwtService.generateToken(userDetails);
+            String token = authService.createToken(loginRequest.getUsername(), loginRequest.getPassword());
 
             response.put("status", "successful");
-            response.put("token", jwtToken);
+            response.put("token", token);
 
             // Return token in response
             return ResponseEntity.ok(response);
+
         } catch (BadCredentialsException e) {
 
             response.put("status", "error");
