+++ src/main/java/se/mow_e/controllers/MyControlous.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class MyControlous {
 
-
     @GetMapping("/hello")
     public String helloWorld() {
         return "Hello World";
     }
+
 }
