+++ src/main/java/se/mow_e/services/AuthService.java
@@ -1,0 +1,34 @@
+package se.mow_e.services;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+@Service
+public class AuthService {
+
+    @Autowired
+    private AuthenticationManager authenticationManager;
+
+    @Autowired
+    private JwtService jwtService;
+
+
+    public String createToken(String username, String password){
+
+        // Authenticate user
+        Authentication authentication = authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(username, password)
+        );
+
+        // Generate JWT token
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+        return jwtService.generateToken(userDetails);
+    }
+
+}
+
